# docs/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Marketing Analytics",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/sites": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "List Dashboard Sites",
        "operationId": "list_dashboard_sites_api_sites_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSitesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/shuffle": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Shuffle Ingest",
        "operationId": "shuffle_ingest_api_shuffle_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShuffleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/site-access": {
      "get": {
        "tags": [
          "site-access"
        ],
        "summary": "List Site Access",
        "operationId": "list_site_access_api_site_access_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteAccessListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "site-access"
        ],
        "summary": "Grant Site Access",
        "operationId": "grant_site_access_api_site_access_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteAccessGrantRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteAccessListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/site-access/{username}": {
      "delete": {
        "tags": [
          "site-access"
        ],
        "summary": "Revoke Site Access",
        "operationId": "revoke_site_access_api_site_access__username__delete",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteAccessListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sdk/bootstrap": {
      "post": {
        "tags": [
          "sdk"
        ],
        "summary": "Sdk Bootstrap",
        "operationId": "sdk_bootstrap_api_sdk_bootstrap_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SdkBootstrapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkBootstrapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sdk/verify-install": {
      "get": {
        "tags": [
          "sdk"
        ],
        "summary": "Sdk Verify Install",
        "operationId": "sdk_verify_install_api_sdk_verify_install_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          },
          {
            "name": "lookback_minutes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1440,
              "minimum": 1,
              "default": 15,
              "title": "Lookback Minutes"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkInstallVerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DashboardSiteSummary": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "site_name": {
            "type": "string",
            "title": "Site Name"
          },
          "allowed_origin": {
            "type": "string",
            "title": "Allowed Origin"
          },
          "plan": {
            "type": "string",
            "enum": [
              "free",
              "standard",
              "pro"
            ],
            "title": "Plan",
            "default": "free"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "site_name",
          "allowed_origin"
        ],
        "title": "DashboardSiteSummary"
      },
      "DashboardSitesResponse": {
        "properties": {
          "sites": {
            "items": {
              "$ref": "#/components/schemas/DashboardSiteSummary"
            },
            "type": "array",
            "title": "Sites"
          }
        },
        "type": "object",
        "required": [
          "sites"
        ],
        "title": "DashboardSitesResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrivatizedEvent": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "uniques",
              "pageviews",
              "sessions",
              "conversions",
              "revenue"
            ],
            "title": "Kind"
          },
          "payload": {
            "type": "object",
            "title": "Payload"
          },
          "epsilon_used": {
            "type": "number",
            "title": "Epsilon Used"
          },
          "sampling_rate": {
            "type": "number",
            "title": "Sampling Rate"
          },
          "client_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Client Timestamp"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "kind",
          "payload",
          "epsilon_used",
          "sampling_rate",
          "client_timestamp"
        ],
        "title": "PrivatizedEvent"
      },
...
      "SdkBootstrapConfig": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "sampling_rate": {
            "type": "number",
            "title": "Sampling Rate"
          },
          "epsilon_budget": {
            "type": "number",
            "title": "Epsilon Budget"
          },
          "shuffle_url": {
            "type": "string",
            "title": "Shuffle Url"
          },
          "token_ttl_seconds": {
            "type": "integer",
            "title": "Token Ttl Seconds"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "sampling_rate",
          "epsilon_budget",
          "shuffle_url",
          "token_ttl_seconds"
        ],
        "title": "SdkBootstrapConfig"
      },
      "SdkBootstrapRequest": {
        "properties": {
          "site_key": {
            "type": "string",
            "minLength": 12,
            "title": "Site Key"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          }
        },
        "type": "object",
        "required": [
          "site_key"
        ],
        "title": "SdkBootstrapRequest"
      },
      "SdkBootstrapResponse": {
        "properties": {
          "upload_token": {
            "type": "string",
            "title": "Upload Token"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "config": {
            "$ref": "#/components/schemas/SdkBootstrapConfig"
          }
        },
        "type": "object",
        "required": [
          "upload_token",
          "expires_at",
          "config"
        ],
        "title": "SdkBootstrapResponse"
      },
      "SdkInstallVerifyResponse": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "lookback_minutes": {
            "type": "integer",
            "title": "Lookback Minutes"
          },
          "has_recent_activity": {
            "type": "boolean",
            "title": "Has Recent Activity"
          },
          "recent_reports": {
            "type": "integer",
            "title": "Recent Reports"
          },
          "counts_by_kind": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts By Kind"
          },
          "last_report_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Report At"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "lookback_minutes",
          "has_recent_activity",
          "recent_reports",
          "counts_by_kind"
        ],
        "title": "SdkInstallVerifyResponse"
      },
      "ShuffleRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "nonce": {
            "type": "string",
            "title": "Nonce"
          },
          "batch": {
            "items": {
              "$ref": "#/components/schemas/PrivatizedEvent"
            },
            "type": "array",
            "title": "Batch"
          }
        },
        "type": "object",
        "required": [
          "token",
          "nonce",
          "batch"
        ],
        "title": "ShuffleRequest"
      },
      "SiteAccessGrantRequest": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username"
          },
          "role": {
            "const": "member",
            "title": "Role",
            "default": "member"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "username"
        ],
        "title": "SiteAccessGrantRequest"
      },
      "SiteAccessListResponse": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/SiteAccessMemberResponse"
            },
            "type": "array",
            "title": "Members"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "members"
        ],
        "title": "SiteAccessListResponse"
      },
      "SiteAccessMemberResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "member"
            ],
            "title": "Role"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "username",
          "role"
        ],
        "title": "SiteAccessMemberResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}